feat: Move Assessment Statement into history
* Combine HistoryCard.tests into a single file
* Add new tests for assessment statement
* default expand first history item
