+++ frontend/src/hooks/useCurrentUser.js
@@ -21,7 +21,9 @@ export const useCurrentUser = () => {
     onSuccess: setUser,
     onError: (error) => {
       console.error('Error fetching current user:', error)
-    }
+    },
+    staleTime: 5 * (60 * 1000), // 5 mins
+    cacheTime: 10 * (60 * 1000) // 10 mins
   })
 
   /**
